+++ controllers/AdminController.php
@@ -3,21 +3,17 @@
 namespace humhub\modules\fcmPush\controllers;
 
 use humhub\modules\admin\components\Controller;
-use humhub\modules\admin\notifications\NewVersionAvailable;
 use humhub\modules\fcmPush\models\FcmUser;
 use humhub\modules\fcmPush\Module;
-use humhub\modules\user\models\User;
 use Yii;
 
 /**
- *
  * @property Module $module
  */
 class AdminController extends Controller
 {
     public function actionIndex()
     {
-
         $model = $this->module->getConfigureForm();
 
         if ($model->load(Yii::$app->request->post()) && $model->validate() && $model->saveSettings()) {
@@ -28,34 +24,24 @@ class AdminController extends Controller
         return $this->render('index', ['model' => $model]);
     }
 
-    /**
-     * @return string
-     */
-    public function actionMobileApp()
+    public function actionDebug()
     {
-        if (Yii::$app->request->get('triggerNotification') == 1) {
-
-            /** @var User $user */
-            $user = Yii::$app->user->getIdentity();
-
-            $updateNotification = new NewVersionAvailable();
-            $updateNotification->sendBulk(User::find()->where(['user.id' => $user->id]));
-            $this->view->setStatusMessage('success', 'Notification queued!');
-            return $this->redirect('mobile-app');
-        }
-
-        if (Yii::$app->request->get('deleteToken') != "") {
-            $t = FcmUser::findOne(['id' => Yii::$app->request->get('deleteToken')]);
-            if ($t->delete() !== false) {
-                $this->view->setStatusMessage('success', 'Token deleted!');
-                return $this->redirect('mobile-app');
+        if ($tokenId = Yii::$app->request->get('deleteToken')) {
+            $token = FcmUser::findOne(['id' => $tokenId]);
+            if ($token->delete()) {
+                $this->view->success('Token deleted!');
+            } else {
+                $this->view->warn('Token NOT deleted!');
             }
-
-            $this->view->setStatusMessage('warning', 'Token NOT deleted!');
-            return $this->redirect('mobile-app');
+            return $this->redirect('index');
         }
 
-        return $this->render('mobile-app');
+        return $this->renderAjax('debug', [
+            'tokens' => FcmUser::find()
+                ->where(['user_id' => Yii::$app->user->id])
+                ->orderBy('created_at DESC')
+                ->all(),
+        ]);
     }
 
 }